Move exchange info to market_response
